edited 0

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -8,7 +8,7 @@
  */
 size_t print_listint(const listint_t *h)
 {
-	int count = 0;
+	size_t count = 0;
 
 	while (h)
 	{
@@ -16,7 +16,6 @@ size_t print_listint(const listint_t *h)
 		h = h->next;
 		count++;
 	}
-
 	return (count);
 }
